[DEVELOPMENT ENVIRONMENT] * BUG: Fix version requirements in packager. * BUG: Makes autotools packager independant of the Gambas version. * BUG: Don't use the release number if not specified in the autotools   packager. * BUG: The autotools packager now use the compilation flags defined in the   project properties * NEW: The packager can make packages for component projects. Ubuntu and   Debian are not supported yet.
[64 BITS]
* BUG: Fixed some uses of long datatype in the gb.qt.ext component, the 
  Splitter and the Container class of gb.qt.


git-svn-id: svn://localhost/gambas/branches/2.0@1029 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.qt/src/CContainer.cpp
@@ -188,7 +188,7 @@ static void arrange_later(QWidget *cont)
   GB.Ref(_object);
   //qDebug("later: %p: dirty = TRUE", THIS);
   THIS_ARRANGEMENT->dirty = TRUE;
-  GB.Post((void (*)())post_arrange_later, (long)THIS);
+  GB.Post((void (*)())post_arrange_later, (intptr_t)THIS);
 }
 #endif
